+++ src/EnemyCar.cpp
@@ -1,6 +1,6 @@
 #include "../Include/EnemyCar.h"
 
-EnemyCar::EnemyCar(int x, int y, int carType, Direction moveDirection) : Car("car" + std::to_string(carType)){
+EnemyCar::EnemyCar(float x, float y, int carType, Direction moveDirection) : Car("car" + std::to_string(carType)){
     this->carType = carType;
     this->x = x;
     this->y = y;
@@ -39,11 +39,11 @@ void EnemyCar::render(){
     Car::render(x, y);
 }
 
-bool EnemyCar::futureCarCollider(EnemyCar* otherCar, int time){
-    int addX = moveDirection == LEFT ? -time : (moveDirection == RIGHT ? time : 0);
-    int addOtherCarX = otherCar->moveDirection == LEFT ? -time : (otherCar->moveDirection == RIGHT ? time : 0);
-    int addY = moveDirection == UP ? -time : (moveDirection == DOWN ? time : 0);
-    int addOtherCarY = otherCar->moveDirection == UP ? -time : (otherCar->moveDirection == DOWN ? time : 0);
+bool EnemyCar::futureCarCollider(EnemyCar* otherCar, float time){
+    float addX = moveDirection == LEFT ? -time : (moveDirection == RIGHT ? time : 0);
+    float addOtherCarX = otherCar->moveDirection == LEFT ? -time : (otherCar->moveDirection == RIGHT ? time : 0);
+    float addY = moveDirection == UP ? -time : (moveDirection == DOWN ? time : 0);
+    float addOtherCarY = otherCar->moveDirection == UP ? -time : (otherCar->moveDirection == DOWN ? time : 0);
 
     if(x + addX > otherCar->x + otherCar->w + addOtherCarX || x + w + addX < otherCar->x + addOtherCarX) return false;
     if(y + addY > otherCar->y + otherCar->h + addOtherCarY || y + h + addY < otherCar->y + addOtherCarY) return false;
